refactor: refine the operation.py name to prevent from dep name conflict
